requirements.txt
numpy
opencv-python
scikit-image
pyyaml
imageio